Initial Commit
I've proceeded through Section 6.1 of the Agile Web Development with
Rails book (4th ed.), and have created the Rails Application for my
Depot app.  I've created the database, generated the scaffolding, and
applied the migration as per the instructions.  I've also created a
single product so as to populate my Products listing with something.